Fix for Java long.
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1249 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/xmlvm2objc/compat-lib/objc/xmlvm.m
@@ -109,7 +109,7 @@ void xmlvm_init()
        sizeOfBaseType = sizeof(double);
        break;
     case 8: // long
-       sizeOfBaseType = sizeof(long);
+       sizeOfBaseType = sizeof(long long);
        break;
     default: // object reference
        sizeOfBaseType = sizeof(id);
